requirements.txt
crewai==0.28.0
langchain==0.1.0
langchain-openai==0.0.5
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.4
numpy==1.24.3
yfinance==0.2.28
newsapi-python==0.2.6
openai==1.3.7
python-dateutil==2.8.2
schedule==1.2.0
logging==0.4.9.6 